recognition/requirements.txt
absl-py>=2.1.0
astunparse>=1.6.3
certifi>=2024.2.2
charset-normalizer>=3.3.2
colorama>=0.4.6
confluent-kafka>=2.3.0
contourpy>=1.2.1
cycler>=0.12.1
filelock>=3.14.0
flatbuffers>=24.3.25
fonttools>=4.51.0
fsspec>=2024.3.1
gast>=0.5.4
gitdb>=4.0.11
GitPython>=3.1.43
google-pasta>=0.2.0
grpcio>=1.63.0
h5py>=3.11.0
idna>=3.7
importlib_metadata>=7.1.0
importlib_resources>=6.4.0
intel-openmp>=2021.4.0
Jinja2>=3.1.3
joblib>=1.4.2
keras>=3.3.3
keras-facenet>=0.3.2
kiwisolver>=1.4.5
libclang>=18.1.1
Markdown>=3.6
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
matplotlib>=3.8.4
mdurl>=0.1.2
mkl>=2021.4.0
ml-dtypes>=0.3.2
mpmath>=1.3.0
mtcnn>=0.1.1
namex>=0.0.8
networkx>=3.2.1
numpy>=1.26.4
opencv-python>=4.9.0.80
opt-einsum>=3.3.0
optree>=0.11.0
packaging>=24.0
paho-mqtt>=2.1.0
pandas>=2.2.2
pillow>=10.3.0
protobuf>=4.25.3
psutil>=5.9.8
py-cpuinfo>=9.0.0
Pygments>=2.17.2
pyparsing>=3.1.2
python-dateutil>=2.9.0.post0
pytz>=2024.1
PyYAML>=6.0.1
requests>=2.32.3
rich>=13.7.1
scikit-learn>=1.4.2
scipy>=1.13.0
seaborn>=0.13.2
six>=1.16.0
smmap>=5.0.1
sympy>=1.12
tbb>=2021.12.0
tensorboard>=2.16.2
tensorboard-data-server>=0.7.2
tensorflow>=2.16.1
tensorflow-intel>=2.16.1
tensorflow-io-gcs-filesystem>=0.31.0
termcolor>=2.4.0
thop>=0.1.1.post2209072238
threadpoolctl>=3.5.0
torch>=2.3.0
torchvision>=0.18.0
tqdm>=4.66.4
typing_extensions>=4.11.0
tzdata>=2024.1
ultralytics>=8.2.8
urllib3>=2.2.1
Werkzeug>=3.0.2
wrapt>=1.16.0
zipp>=3.18.1
